Split jaws.go session and broadcast clusters into their own files
Pure code-motion, no behavior change: build, go vet, staticcheck,
golangci-lint and the full -race / -tags debug -race suites are unchanged.

- Move the Jaws-side session helpers (SessionCount, Sessions, getSessionLocked,
  getCookieSessionsIDs, GetSession, NewSession, newSession, deleteSession) and
  the session middleware (sessioner, Jaws.Session) into session.go, which
  already owns the Session methods.
- Move the broadcast/dirt/redirect/alert machinery and the SetInner..JsCall
  convenience wrappers into a new broadcast.go.

jaws.go shrinks from 1395 to 936 lines and now holds the instance lifecycle,
request pool, Serve loop, client-IP helpers and HTTP routing.
